[refactor] Migrate PaymentForm to separated component
