+++ src/pages/Payment/PaymentForm/PaymentForm.module.scss
@@ -1,0 +1,91 @@
+.root {
+  margin: 0 1em;
+  .summary_info,
+  .billing_address {
+    display: flex;
+    flex-direction: column;
+    gap: 0.8em;
+    border: 1px solid hsla(0, 0%, 90%, 1);
+    border-radius: 6px;
+    padding: 1em 1.2em;
+
+    .summary_contact,
+    .summary_shipping,
+    .summary_method {
+      display: flex;
+      align-items: center;
+    }
+
+    p {
+      font-size: 0.9rem;
+      color: var(--text-secondary);
+      margin-right: 0.5em;
+      font-weight: 500;
+    }
+
+    span {
+      font-size: 0.9rem;
+    }
+  }
+
+  .line {
+    content: "";
+    background-color: hsla(0, 0%, 90%, 1);
+    height: 1px;
+    width: 100%;
+  }
+
+  .payment {
+    margin-top: 2em;
+
+    input {
+      width: 100%;
+      padding: 0.8em 1.2em;
+      border-radius: 0;
+      border: 1px solid hsla(0, 0%, 90%, 1);
+      font-size: 0.9rem;
+      color: var(--text-secondary);
+
+      &:focus {
+        outline: 1px solid var(--text-tertiary);
+      }
+    }
+
+    .btn_container {
+      margin-top: 2em;
+    }
+
+    .card_details {
+      margin: 2em 0;
+      display: flex;
+      flex-direction: column;
+      gap: 1em;
+
+      .divide {
+        display: flex;
+        gap: 1em;
+      }
+    }
+
+    .billing_address {
+      margin-top: 1.5em;
+
+      .billing_item {
+        display: flex;
+        align-items: center;
+
+        input {
+          height: 1.2em;
+          width: 1.2em;
+          margin-right: 1em;
+        }
+      }
+    }
+  }
+
+  .error {
+    color: var(--text-error);
+    font-size: 0.8rem;
+    margin-top: 0.5em;
+  }
+}
